Fix regex in reply.rb
There is a regex in reply.rb that is used to match the escape sequences
in the service names.  The digits are extracted and converted to
integers, and then into characters with 'chr'.

The problem is that the regex is too greedy, and will match more than 3
digits, meaning the a subsequent digit that is not part of the escape
code will be consumed.  The is incorrect, but also causes an exception
if the number is larger that what chr would expect.

This patch fixes the issue (and includes a test for set_full_name)